fix scripts

+++ build/prebuild.sh
@@ -9,6 +9,8 @@ fi
 #if [ -z "$OML_APIKEY" ]; then echo "Environment variable OML_APIKEY is unset or empty." && exit 1; fi
 #if [ -z "$TWINE_PASSWORD" ]; then echo "Environment variable TWINE_PASSWORD is unset or empty." && exit 1; fi
 
-rm dist/* -r
+if test -d "dist"; then
+	rm dist/* -r
+fi
 
 echo "build starting..."
